Clean up MSBuild integration tests
Remove redundant test cases and merge all tests into a single file
