+++ src/MdDocs.MSBuild.IntegrationTest/_Helpers/MSBuildRuntimeInfo.cs
