+++ src/MdDocs.MSBuild.IntegrationTest/_Helpers/MSBuildRuntimeType.cs
