+++ src/MdDocs.MSBuild.IntegrationTest/_Helpers/VSWhereException.cs
